requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
albucore==0.0.13
albumentations==1.4.10
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
async-timeout==5.0.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dacite==1.8.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
imageio==2.36.1
imgaug==0.4.0
inexactsearch==1.0.2
inference-sdk==0.31.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
mypy-extensions==1.0.0
narwhals==1.18.4
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==3.0.0b1
pandas==2.2.3
patsy==1.0.1
phik==0.12.4
pillow==10.4.0
plotly==5.24.1
propcache==0.2.1
protobuf==5.29.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
silpa_common==0.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soundex==1.1.3
soupsieve==2.6
SQLAlchemy==2.0.36
statsmodels==0.14.4
streamlit==1.41.1
supervision==0.25.1
sympy==1.13.3
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
toml==0.10.2
tomli==2.2.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
typeguard==4.4.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visions==0.7.6
watchdog==6.0.0
Werkzeug==3.1.3
wordcloud==1.9.4
yarl==1.18.3
ydata-profiling==4.12.1
